Move metadata.
