+++ python/lib/sift_client/util/metadata.py
